Add files via upload
The file blob_commands.txt contains the pipeline for obtaining all the Midichloria genomes of Floriano et al. biorXiv 2022 (Blobology pipeline by Kumar et al. 2015, assembly graph visualization with Bandage, removal of short contigs) except for the Midichloria symbiont of I. holocyclus, for which the pipeline is presented in Ihol_blob_commands.txt. 
The file blobparameters.tsv contains the parameters used for each organism for the Blobology selection on GC content and coverage.
The directory plots contains all the plots obtained with the Blobology pipeline.
filter_SPAdes_contigs.py is a simple script to filter SPAdes assemblies on the basis of contig length, coverage and GC content estimated by SPAdes. In our analyses, this was used only for selection on the basis of contig length (last step).